tests/wpt/tests/css/css-sizing/responsive-iframe/responsive-iframe-logical.tentative.html

<!doctype HTML>
<title>Test that `contain-intrinsic-block-size` makes IFRAMEs be responsively sized.</title>
<link rel="author" href="mailto:kojii@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-sizing/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
iframe {
  border: none;
}

#block {
  contain-intrinsic-block-size: from-element none;
}

#inline {
  contain-intrinsic-inline-size: from-element none;
}
</style>
<iframe id="block" frameborder=0 src="resources/iframe-contents.html"></iframe>
<iframe id="inline" frameborder=0 src="resources/iframe-contents.html"></iframe>
<script>
async_test(t => {
  window.addEventListener('load', t.step_func_done(e => {
    const block = document.getElementById('block');
    assert_equals(block.offsetHeight, 400);
    const inline = document.getElementById('inline');
    assert_equals(inline.offsetHeight, 150);
  }));
}, "`contain-intrinsic-block-size` makes IFRAMEs be responsively sized");
</script>
